opa: run more tests on OPA, not EOPA
Also skip the EOPA-dependent tests when the license env var isn't set.

This should allow us to have more confidence into the dependabot-PRs.
They don't have access to the repository secrets, so they wouldn't pass
before. Now, they can run a large subset of the test suite and we'll
skip the tests that depend on EOPA.

Signed-off-by: Stephan Renatus <stephan@styra.com>