opa-react: enumerate batch items, adjust tests to not depend on hashes
This way, our payloads sent to the Batch API become a lot smaller, and less
redundant. The WeakMap should ensure that our memory usage doesn't grow too
much from the extra accounting.

More thought needs to go into the choice of "counter". It doesn't have to
outlive the lifetime of an eval batch, but I have yet to find a way to tie
it to that.