package(ucast-prisma): initial commit
This is a package for translating ucast conditions to Prisma filters.
It's in early development, and shouldn't be depended upon just yet.

For this reason, it's not yet published to npm.

Users of this package will need to resort to a service like gitpkg to
install it. Here's an example of how to do that:

    npm install 'https://gitpkg.vercel.app/StyraInc/opa-typescript/packages/ucast-prisma?main&scripts.postinstall=npm%20install%20--ignore-scripts%20%26%26%20npm%20run%20build"'